Create Performing calculations on stock prices
The programme takes the given monthly stock prices of the Amazon and ebay and performs the following calculations: rate of return, variance, standard deviation and correlation to asses the risk in investment into the two companies. The code for the calculations is found in the file: equations for calcs on stock prices.